design: 마이페이지 내정보 css 수정
