+++ src/components/mypage/MyMoimBox.tsx
@@ -7,7 +7,7 @@ type BoxProps = {
 
 const MyMoimBox = ({ title, children }: BoxProps) => {
   return (
-    <section className="flex h-[75vh] flex-col gap-2 rounded-2xl bg-[#E1F0D3] p-8 shadow-sm">
+    <section className="relative flex h-[75vh] flex-col gap-2 rounded-2xl bg-[#E1F0D3] p-8 shadow-sm">
       {title && (
         <h2 className="mb-8 text-center text-2xl font-semibold">{title}</h2>
       )}
