+++ src/mockup/user.json
@@ -9,5 +9,6 @@
   "address": "서울시 강남구",
   "intro": "안녕하세요! 열심히 공부 중입니다.",
   "github_url": "https://github.com/woojung",
-  "portfolio_url": "https://woojung.dev"
+  "portfolio_url": "https://woojung.dev",
+  "position_name": "프론트엔드"
 }
